hw04 comment fixes
