+++ hw04/src/main/java/ru/otus/hw/dao/InMemoryResultDaoImpl.java
@@ -2,6 +2,7 @@ package ru.otus.hw.dao;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Primary;
+import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
 import ru.otus.hw.domain.Student;
 import ru.otus.hw.domain.TestResult;
@@ -10,17 +11,16 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Stream;
 
 import static java.util.Optional.ofNullable;
 
-@Service
+@Repository
 @RequiredArgsConstructor
-@Primary
 public class InMemoryResultDaoImpl implements ResultDao {
 
-    //Т.к. конкурентного доступа не подразумевается, используем обычную мапу
-    private final Map<Student, List<TestResult>> resultMap = new HashMap<>();
+    private final Map<Student, List<TestResult>> resultMap = new ConcurrentHashMap<>();
 
     @Override
     public void save(TestResult testResult) {
