+++ hw04/src/main/java/ru/otus/hw/service/TestRunnerService.java
@@ -1,0 +1,13 @@
+package ru.otus.hw.service;
+
+import ru.otus.hw.domain.TestResult;
+
+import java.util.List;
+
+public interface TestRunnerService {
+    void run();
+
+    List<TestResult> getAllPreviousResults();
+
+    List<TestResult> getPreviousResultsForStudent(String firstName, String lastName);
+}
